v2/requirements.txt
fastapi
uvicorn
python-multipart
lancedb
#older ones
litellm
pymupdf4llm
spacy
matplotlib
fixit
transformers
rerankers
torch
pytest-asyncio
sentence_transformers
